Moved edit and delte article tests to a single test file
