+++ .github/workflows/deploy.yaml
@@ -66,7 +66,7 @@ jobs:
         if: always()
         run: |
           cd tests/integration/tests/ui_tests
-          python -m pytest -n 4 --browser headless-chrome -k 'not test_snapshots' --app_url http://localhost:5000
+          python -m pytest -n 2 --browser headless-chrome -k 'not test_snapshots' --app_url http://localhost:5000
 
   deploy-staging:
     name: Deploy to staging
